pyproject.toml
[tool.poetry]
name = "qtest"
version = "0.1.0"
description = ""
authors = ["Sameer Segal <sameersegal@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-storage-queue = "^12.6.0"
applicationinsights = "^0.11.10"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
